requirements.txt
altair==4.2.0
plotly==5.9.0
snowflake_connector_python==3.0.3
streamlit==1.22.0
streamlit_option_menu==0.3.2
pyarrow==10.0.1